backend/requirements.txt
Flask
Flask-Cors
boto3
pandas
mysql-connector-python
